Initial support for dynamic data
This commit implements support for clients providing read and size
functions when creating a "virtual" file. When programs try to
read or determine the size of the file, these functions are called and
they can incrementally read from their data source. This is very
useful for platforms with large data sources (e.g., databases whose
tables we want to represent as CSV) or dynamic data sources.